contract_engine/requirements.txt
pyyaml>=6.0
pydantic>=2.0
jsonschema>=4.20
httpx>=0.27
